refactor: split the Raft inherent impl into per-concern files
# Summary

Move the 49-method `impl Raft<C, SM>` out of `raft/mod.rs` into four
sibling files grouped by what the caller is doing, leaving only
construction and the small accessors behind.

# Details

`raft/mod.rs` was 2007 lines, of which one impl block accounted for
1380. Everything the crate exposes to an application lived in that one
block, so finding a method meant scrolling past every unrelated one, and
any edit landed in a file that half the crate already has open.

The methods were already laid out by concern, so each new file is a
contiguous run of the original and keeps the original ordering:

- `impl_raft_protocol.rs` — the RPCs a peer sends this node
- `impl_raft_client.rs` — the application's reads and writes
- `impl_raft_watch.rs` — metrics, progress watchers, and waits
- `impl_raft_lifecycle.rs` — initialize, shutdown, and the closures that
  run against core state or the state machine

This follows the existing `impl_raft_blocking_write.rs` precedent rather
than inventing a new layout.

Movement only: no signature, body, or doc comment changed, so no
`#[since]` is needed and the public API test still passes unchanged.
Each new file carries just the imports its own methods need, and
`mod.rs` drops the ones that left with them.
