+++ openraft/src/raft/mod.rs
@@ -12,6 +12,10 @@ pub(crate) mod api;
 #[cfg(test)]
 mod declare_raft_types_test;
 mod impl_raft_blocking_write;
+mod impl_raft_client;
+mod impl_raft_lifecycle;
+mod impl_raft_protocol;
+mod impl_raft_watch;
 pub mod linearizable_read;
 pub(crate) mod message;
 mod raft_inner;
@@ -29,7 +33,6 @@ pub(in crate::raft) mod core_state;
 mod leader;
 
 use std::fmt::Debug;
-use std::future::Future;
 use std::sync::Arc;
 use std::sync::Mutex;
 use std::time::Duration;
@@ -37,7 +40,6 @@ use std::time::Duration;
 use core_state::CoreState;
 use derive_more::Display;
 use futures_util::FutureExt;
-use linearizable_read::Linearizer;
 pub use message::AppendEntriesRequest;
 pub use message::AppendEntriesResponse;
 pub use message::ClientWriteResponse;
@@ -66,17 +68,12 @@ pub use self::watch_handle::WatchChangeHandle;
 use crate::Extensions;
 use crate::OptionalSend;
 use crate::RaftNetworkFactory;
-use crate::RaftState;
 pub use crate::RaftTypeConfig;
 use crate::StorageError;
 use crate::StorageHelper;
 use crate::async_runtime::MpscWeakSender;
-use crate::async_runtime::OneshotSender;
 use crate::async_runtime::mpsc::MpscSender;
 use crate::async_runtime::watch::WatchReceiver;
-use crate::base::BoxFuture;
-use crate::base::BoxOnce;
-use crate::base::BoxStream;
 use crate::config::Config;
 use crate::config::RuntimeConfig;
 use crate::core::ClientResponderQueue;
@@ -85,16 +82,10 @@ use crate::core::SharedReplicateBatch;
 use crate::core::StepDownWatcher;
 use crate::core::Tick;
 use crate::core::heartbeat::handle::HeartbeatWorkersHandle;
-use crate::core::io_flush_tracking::AppliedProgress;
-use crate::core::io_flush_tracking::CommitProgress;
 pub use crate::core::io_flush_tracking::FlushPoint;
 use crate::core::io_flush_tracking::IoProgressWatcher;
-use crate::core::io_flush_tracking::LogProgress;
-use crate::core::io_flush_tracking::SnapshotProgress;
-use crate::core::io_flush_tracking::VoteProgress;
 use crate::core::merged_raft_msg_receiver::BatchRaftMsgReceiver;
 use crate::core::notification::Notification;
-use crate::core::raft_msg::RaftMsg;
 use crate::core::raft_msg::external_command::ExternalCommand;
 use crate::core::raft_msg::install_full_snapshot_request::InstallFullSnapshotRequest;
 use crate::core::runtime_stats::RuntimeStats;
@@ -102,21 +93,12 @@ use crate::core::sm;
 use crate::core::sm::worker;
 use crate::engine::Engine;
 use crate::engine::EngineConfig;
-use crate::entry::EntryPayload;
-use crate::errors::ClientWriteError;
 use crate::errors::Fatal;
 use crate::errors::ForwardToLeader;
-use crate::errors::InitializeError;
-use crate::errors::LinearizableReadError;
-use crate::errors::RaftError;
-use crate::errors::into_raft_result::IntoRaftResult;
-use crate::membership::IntoNodes;
 use crate::metrics::MetricsRecorder;
 use crate::metrics::RaftDataMetrics;
 use crate::metrics::RaftMetrics;
 use crate::metrics::RaftServerMetrics;
-use crate::metrics::Wait;
-use crate::metrics::WaitError;
 use crate::network::NetSnapshot;
 use crate::raft::raft_inner::RaftInner;
 pub use crate::raft::runtime_config_handle::RuntimeConfigHandle;
@@ -125,17 +107,9 @@ use crate::raft_state::IOId;
 use crate::storage::RaftLogStorage;
 use crate::storage::RaftStateMachine;
 use crate::type_config::TypeConfigExt;
-use crate::type_config::alias::JoinErrorOf;
-use crate::type_config::alias::LogIdOf;
 use crate::type_config::alias::MpscSenderOf;
 use crate::type_config::alias::MpscWeakSenderOf;
-use crate::type_config::alias::NodeIdOf;
-use crate::type_config::alias::SnapshotDataOf;
-use crate::type_config::alias::SnapshotOf;
-use crate::type_config::alias::VoteOf;
 use crate::type_config::alias::WatchReceiverOf;
-use crate::type_config::alias::WriteResponderOf;
-use crate::vote::Vote;
 use crate::vote::leader_id::raft_leader_id::RaftLeaderId;
 use crate::vote::leader_id::raft_leader_id::RaftLeaderIdExt;
 use crate::vote::non_committed::UncommittedVote;
@@ -889,1119 +863,5 @@ where
     /// Returns [`Fatal`] error if RaftCore is shut down or has a storage error.
     pub async fn set_metrics_recorder(&self, recorder: Option<Arc<dyn MetricsRecorder>>) -> Result<(), Fatal<C>> {
         self.inner.send_external_command(ExternalCommand::SetMetricsRecorder { recorder }).await
-    }
-
-    /// Submit an AppendEntries RPC to this Raft node.
-    ///
-    /// These RPCs are sent by the cluster leader to replicate log entries (§5.3), and are also
-    /// used as heartbeats (§5.2).
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub async fn append_entries(&self, rpc: AppendEntriesRequest<C>) -> Result<AppendEntriesResponse<C>, RaftError<C>> {
-        self.protocol_api().append_entries(rpc).await.into_raft_result()
-    }
-
-    /// Submit a stream of AppendEntries RPCs to this Raft node.
-    ///
-    /// This is a stream-oriented version of [`Self::append_entries`] with pipelining support.
-    /// It spawns a background task that reads from the input stream, sends requests to RaftCore,
-    /// and forwards response receivers to the output stream. Responses are yielded in order.
-    ///
-    /// ## Pipelining Behavior
-    ///
-    /// - A background task reads from the input stream and sends to RaftCore
-    /// - Uses a bounded channel (64 slots) for backpressure
-    /// - Responses are yielded in order (FIFO) as they complete
-    ///
-    /// ## Output
-    ///
-    /// The output stream emits:
-    /// - `Ok(log_id)` when logs are successfully flushed
-    /// - `Err(e)` when an error occurs, which terminates the stream
-    ///
-    /// ## Pinning
-    ///
-    /// The returned stream is `!Unpin` because it uses async closures internally.
-    /// You must pin the stream before calling `.next()`:
-    ///
-    /// ```ignore
-    /// use std::pin::pin;
-    ///
-    /// let mut output = pin!(raft.stream_append(input));
-    /// while let Some(result) = output.next().await { /* ... */ }
-    /// ```
-    ///
-    /// Alternatively, use `Box::pin` for heap pinning if the stream needs to be stored or returned:
-    ///
-    /// ```ignore
-    /// let mut output = Box::pin(raft.stream_append(input));
-    /// ```
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// use std::pin::pin;
-    /// use futures_util::StreamExt;
-    ///
-    /// let input_stream = futures_util::stream::iter(vec![request1, request2, request3]);
-    /// let mut output_stream = pin!(raft.stream_append(input_stream));
-    ///
-    /// while let Some(result) = output_stream.next().await {
-    ///     match result {
-    ///         Ok(Ok(log_id)) => println!("Flushed: {:?}", log_id),
-    ///         Ok(Err(err)) => {
-    ///             println!("Append error: {}", err);
-    ///             break;
-    ///         }
-    ///         Err(fatal) => {
-    ///             println!("Fatal: {}", fatal);
-    ///             break;
-    ///         }
-    ///     }
-    /// }
-    /// ```
-    #[since(version = "0.10.0", change = "stream item contains Fatal")]
-    #[since(version = "0.10.0")]
-    pub fn stream_append<S>(
-        &self,
-        stream: S,
-    ) -> impl futures_util::Stream<Item = Result<StreamAppendResult<C>, Fatal<C>>> + OptionalSend + 'static
-    where
-        S: futures_util::Stream<Item = AppendEntriesRequest<C>> + OptionalSend + 'static,
-    {
-        self.protocol_api().stream_append(stream)
-    }
-
-    /// Submit a VoteRequest (RequestVote in the spec) RPC to this Raft node.
-    ///
-    /// These RPCs are sent by cluster peers which are in candidate state attempting to gather votes
-    /// (§5.2).
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub async fn vote(&self, rpc: VoteRequest<C>) -> Result<VoteResponse<C>, RaftError<C>> {
-        self.protocol_api().vote(rpc).await.into_raft_result()
-    }
-
-    /// Submit a Pre-Vote RPC to this Raft node.
-    ///
-    /// A pre-candidate sends this before incrementing its term to ask whether peers *would* grant
-    /// it a vote at `rpc.vote` (a hypothetical next term). Handling this RPC never persists a
-    /// vote or changes this node's term; it only reports whether the vote would be granted,
-    /// judged by the same leader-lease and last-log-id rules as [`Raft::vote()`]. Pre-Vote is
-    /// only used when [`Config::enable_pre_vote`](crate::Config::enable_pre_vote) is enabled on
-    /// the sender.
-    #[since(version = "0.10.0", change = "added for the Pre-Vote feature")]
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub async fn pre_vote(&self, rpc: VoteRequest<C>) -> Result<VoteResponse<C>, RaftError<C>> {
-        self.protocol_api().pre_vote(rpc).await.into_raft_result()
-    }
-
-    /// Get the latest snapshot from the state machine.
-    ///
-    /// The request is served directly by the state-machine worker, not `RaftCore`. It returns an
-    /// error only when that worker fails to serve it, e.g., encountering a storage error or having
-    /// stopped (`Fatal::Stopped`), which can occur while `RaftCore` is still running.
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub async fn get_snapshot(&self) -> Result<Option<SnapshotOf<C, SM::SnapshotData>>, RaftError<C>> {
-        self.protocol_api().get_snapshot().await.into_raft_result()
-    }
-
-    /// Get a snapshot data for receiving snapshot from the leader.
-    ///
-    /// It does not check `Vote` because it is a read operation and does not break raft
-    /// protocol.
-    #[since(version = "0.10.0", change = "SnapshotData without Box")]
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub async fn begin_receiving_snapshot(&self) -> Result<SnapshotDataOf<C, SM>, RaftError<C>> {
-        self.protocol_api().begin_receiving_snapshot().await.into_raft_result()
-    }
-
-    /// Install a completely received snapshot to the state machine.
-    ///
-    /// This method is used to implement an application defined snapshot transmission.
-    /// The application receives a snapshot from the leader, in chunks or a stream, and
-    /// then rebuild a snapshot, then pass the snapshot to Raft to install.
-    #[since(version = "0.9.0")]
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub async fn install_full_snapshot(
-        &self,
-        vote: VoteOf<C>,
-        snapshot: SnapshotOf<C, SM::SnapshotData>,
-    ) -> Result<SnapshotResponse<C>, Fatal<C>> {
-        self.protocol_api().install_full_snapshot(vote, snapshot).await
-    }
-
-    /// Get the ID of the current leader from this Raft node.
-    ///
-    /// This method is based on the Raft metrics system which does a good job at staying
-    /// up-to-date; however, the `is_leader` method must still be used to guard against stale
-    /// reads. This method is perfect for making decisions on where to route client requests.
-    #[tracing::instrument(level = "debug", skip(self))]
-    pub async fn current_leader(&self) -> Option<C::NodeId> {
-        self.metrics().borrow_watched().current_leader.clone()
-    }
-
-    /// Ensures reads performed after this method are linearizable across the cluster
-    /// using an explicitly provided policy. This method is just a shorthand for calling
-    /// [`get_read_log_id()`](Raft::get_read_log_id) and then calling [Raft::wait].
-    ///
-    /// This method is just a shorthand for combining calling
-    /// [`Raft::get_read_linearizer()`](Self::get_read_linearizer) and
-    /// [`Linearizer::try_await_ready()`](Linearizer::try_await_ready), i.e., it is
-    /// equivalent to:
-    /// ```ignore
-    /// my_raft.get_read_linearizer(read_policy).await?.try_await_ready(&my_raft, None).await?;
-    /// ```
-    ///
-    /// To support follower read, i.e., get `read_log_id` on a remote leader then read on local
-    /// state machine, see [`Raft::get_read_linearizer`].
-    ///
-    /// The `read_policy` defines the policy to ensure leadership. See: [`ReadPolicy`].
-    ///
-    /// Returns:
-    /// - `Ok(read_log_id)` on successful confirmation that the node is the leader. `read_log_id`
-    ///   represents the log id up to which the state machine has applied to ensure a linearizable
-    ///   read.
-    /// - `Err(RaftError<LinearizableReadError>)` if fails to assert leadership.
-    ///
-    /// # Examples
-    /// ```ignore
-    /// // Use a strict policy for this specific critical read
-    /// my_raft.ensure_linearizable(ReadPolicy::ReadIndex).await?;
-    ///
-    /// // Or use a more performant policy when consistency requirements are less strict
-    /// my_raft.ensure_linearizable(ReadPolicy::LeaseRead).await?;
-    ///
-    /// // Then proceed with the state machine read
-    /// ```
-    /// Read more about how it works: [Read Operation](crate::docs::protocol::read)
-    #[since(version = "0.9.0")]
-    #[tracing::instrument(level = "debug", skip(self))]
-    pub async fn ensure_linearizable(
-        &self,
-        read_policy: ReadPolicy,
-    ) -> Result<Option<LogIdOf<C>>, RaftError<C, LinearizableReadError<C>>> {
-        let linearizer = self.app_api().get_read_linearizer(read_policy).await.into_raft_result()?;
-
-        // Safe unwrap: it never times out.
-        let state = linearizer.await_ready(self).await?;
-        Ok(Some(state.read_log_id().clone()))
-    }
-
-    /// Legacy method that returns log IDs directly. Use
-    /// [`Raft::get_read_linearizer`] instead.
-    ///
-    /// This method extracts log IDs from a [`Linearizer`] and returns them as a tuple.
-    /// **For new code, use [`Raft::get_read_linearizer`]** which provides a better API.
-    ///
-    /// See [`Raft::get_read_linearizer`] for full documentation.
-    #[since(version = "0.9.0")]
-    #[tracing::instrument(level = "debug", skip(self))]
-    pub async fn get_read_log_id(
-        &self,
-        read_policy: ReadPolicy,
-    ) -> Result<(Option<LogIdOf<C>>, Option<LogIdOf<C>>), RaftError<C, LinearizableReadError<C>>> {
-        let linearizer = self.app_api().get_read_linearizer(read_policy).await.into_raft_result()?;
-
-        let read_log_id = linearizer.read_log_id();
-        let applied = linearizer.applied();
-
-        Ok((Some(read_log_id.clone()), applied.cloned()))
-    }
-
-    /// Ensures this node is leader and returns a [`Linearizer`] to linearize reads.
-    ///
-    /// This method confirms leadership and provides the necessary information to linearize reads
-    /// across the cluster. The leadership is ensured by sending heartbeats or by lease according
-    /// to the specified policy. See: [`ReadPolicy`].
-    ///
-    /// Returns:
-    /// - `Ok(Linearizer<C>)` on successful confirmation that the node is the leader. The
-    ///   [`Linearizer`] contains the `read_log_id` up to which the state machine should apply to
-    ///   linearize reads, and the last `applied` log id.
-    /// - `Err(RaftError<LinearizableReadError>)` if this node fails to ensure its leadership, for
-    ///   example, it detects a higher term, or fails to communicate with a quorum.
-    ///
-    /// Once returned, the caller should block until the state machine to apply up to `read_log_id`
-    /// using [`Linearizer::try_await_ready`].
-    ///
-    /// # Examples
-    /// ```ignore
-    /// let linearizer = my_raft.get_read_linearizer(ReadPolicy::ReadIndex).await?;
-    /// let _ = linearizer.try_await_ready(&my_raft, None).await?.unwrap();
-    ///
-    /// // Following read from state machine is linearized across the cluster
-    /// let val = my_raft.with_state_machine(|sm| { sm.read("foo") }).await?;
-    /// ```
-    ///
-    /// # Follower Read
-    ///
-    /// For follower reads, obtain the `read_log_id` from the leader via application-defined RPC,
-    /// then use [`Linearizer::try_await_ready`] to wait for local state machine to catch
-    /// up.
-    ///
-    /// ```ignore
-    /// // Application defined RPC to get the `read_log_id` from the remote leader
-    /// let leader_id = my_raft.current_leader().await?.unwrap();
-    /// let linearizer = my_app_rpc.get_read_linearizer(leader_id, ReadPolicy::ReadIndex).await?;
-    ///
-    /// // Block waiting local state machine to apply up to to the `read_log_id`
-    /// let _ = linearizer.try_await_ready(&my_raft, None).await?.unwrap();
-    ///
-    /// // Following read from state machine is linearized across the cluster
-    /// let val = my_raft.with_state_machine(|sm| { sm.read("foo") }).await?;
-    /// ```
-    ///
-    /// See: [Read Operation](crate::docs::protocol::read)
-    #[since(version = "0.10.0")]
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub async fn get_read_linearizer(
-        &self,
-        read_policy: ReadPolicy,
-    ) -> Result<Linearizer<C>, RaftError<C, LinearizableReadError<C>>> {
-        self.app_api().get_read_linearizer(read_policy).await.into_raft_result()
-    }
-
-    /// Submit a mutating client request to Raft to update the state of the system (§5.1).
-    ///
-    /// It will be appended to the log, committed to the cluster, and then applied to the
-    /// application state machine. The result of applying the request to the state machine will
-    /// be returned as the response from this method.
-    ///
-    /// Our goal for Raft is to implement linearizable semantics. If the leader crashes after
-    /// committing a log entry but before responding to the client, the client may retry the
-    /// command with a new leader, causing it to be executed a second time. As such, clients
-    /// should assign unique serial numbers to every command. Then, the state machine should
-    /// track the latest serial number processed for each client, along with the associated
-    /// response. If it receives a command whose serial number has already been executed, it
-    /// responds immediately without re-executing the request (§8). The
-    /// [`RaftStateMachine::apply`] method is the perfect place to implement
-    /// this.
-    ///
-    /// These are application specific requirements, and must be implemented by the application
-    /// which is being built on top of Raft.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// // Submit a write request
-    /// let request = MyAppData { key: "foo".to_string(), value: "bar".to_string() };
-    /// let response = raft.client_write(request).await?;
-    /// println!("Applied at log index: {:?}", response.log_id);
-    /// ```
-    #[tracing::instrument(level = "debug", skip(self, app_data))]
-    pub async fn client_write(
-        &self,
-        app_data: C::D,
-    ) -> Result<ClientWriteResponse<C>, RaftError<C, ClientWriteError<C>>> {
-        self.app_api().client_write(EntryPayload::Normal(app_data)).await.into_raft_result()
-    }
-
-    /// Write a blank log entry to the Raft log.
-    ///
-    /// A blank entry contains no application data and is typically used to:
-    /// - Commit entries from previous terms when a new leader is elected
-    /// - Advance the commit index without any state machine changes
-    /// - Act as a barrier to ensure all previous entries are committed
-    ///
-    /// Returns when the blank entry has been applied to the state machine.
-    #[since(version = "0.10.0")]
-    #[tracing::instrument(level = "debug", skip(self))]
-    pub async fn write_blank(&self) -> Result<ClientWriteResponse<C>, RaftError<C, ClientWriteError<C>>> {
-        self.app_api().client_write(EntryPayload::Blank).await.into_raft_result()
-    }
-
-    /// Submit a mutating client request to Raft to update the state machine, returns an application
-    /// defined response receiver [`Responder::Receiver`].
-    ///
-    /// `_ff` means fire and forget.
-    ///
-    /// It is same as [`Self::client_write`] but does not wait for the response.
-    #[since(version = "0.10.0", date = "2025-10-27", change = "add responder arg")]
-    #[since(version = "0.10.0")]
-    pub async fn client_write_ff(
-        &self,
-        app_data: C::D,
-        responder: Option<WriteResponderOf<C>>,
-    ) -> Result<(), Fatal<C>> {
-        self.app_api().client_write_ff(EntryPayload::Normal(app_data), responder).await
-    }
-
-    /// Write multiple application data payloads in a single batch.
-    ///
-    /// Returns a stream that yields each result in submission order.
-    /// This is more efficient than calling [`client_write()`](Self::client_write) multiple times
-    /// as it sends all payloads in a single message to the Raft core.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// use futures_util::TryStreamExt;
-    ///
-    /// let mut stream = raft.client_write_many([data1, data2, data3]).await?;
-    ///
-    /// // try_next() extracts Fatal error, result is WriteResult
-    /// while let Some(result) = stream.try_next().await? {
-    ///     match result {
-    ///         Ok(response) => println!("Applied at log index: {:?}", response.log_id),
-    ///         Err(forward_err) => eprintln!("Forward to leader: {:?}", forward_err),
-    ///     }
-    /// }
-    /// ```
-    #[since(version = "0.10.0")]
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub async fn client_write_many(
-        &self,
-        app_data: impl IntoIterator<Item = C::D>,
-    ) -> Result<BoxStream<'static, Result<WriteResult<C>, Fatal<C>>>, Fatal<C>> {
-        self.app_api().client_write_many(app_data.into_iter().map(EntryPayload::Normal)).await
-    }
-
-    /// Submit a write request to Raft.
-    ///
-    /// Returns a [`WriteRequest`] builder. Fire-and-forget by default;
-    /// use [`.responder()`] for results, [`.with_leader()`] for conditional writes.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// use openraft::impls::ProgressResponder;
-    ///
-    /// // Fire-and-forget
-    /// raft.write(my_data).await?;
-    ///
-    /// // With responder
-    /// let (responder, rx) = ProgressResponder::complete_only();
-    /// raft.write(my_data).responder(responder).await?;
-    /// let result = rx.await??;
-    ///
-    /// // Conditional write (fails if leader changed)
-    /// let leader_id = raft.as_leader()?.to_committed_leader_id();
-    /// raft.write(my_data)
-    ///     .with_leader(leader_id)
-    ///     .responder(responder)
-    ///     .await?;
-    /// ```
-    ///
-    /// [`.responder()`]: WriteRequest::responder
-    /// [`.with_leader()`]: WriteRequest::with_leader
-    #[since(version = "0.10.0")]
-    pub fn write(&self, app_data: C::D) -> WriteRequest<'_, C> {
-        WriteRequest {
-            inner: &self.inner,
-            app_data,
-            responder: None,
-            expected_leader: None,
-        }
-    }
-
-    /// Handle the LeaderTransfer request from a Leader node.
-    ///
-    /// If this node is the `to` node, it resets the Leader lease and triggers an election when the
-    /// expected log entries are flushed.
-    /// Otherwise, it just resets the Leader lease to allow the `to` node to become the Leader.
-    ///
-    /// The application calls
-    /// [`Raft::trigger().transfer_leader()`](crate::raft::trigger::Trigger::transfer_leader) to
-    /// submit Transfer Leader command. Then, the current Leader will broadcast it to every node in
-    /// the cluster via [`RaftNetworkV2::transfer_leader`] and the implementation on the remote node
-    /// responds to transfer leader request by calling this method.
-    ///
-    /// [`RaftNetworkV2::transfer_leader`]: crate::network::RaftNetworkV2::transfer_leader
-    #[since(version = "0.10.0", change = "returns TransferLeaderResponse")]
-    #[since(version = "0.10.0")]
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub async fn handle_transfer_leader(
-        &self,
-        req: TransferLeaderRequest<C>,
-    ) -> Result<TransferLeaderResponse<C>, Fatal<C>> {
-        self.protocol_api().handle_transfer_leader(req).await
-    }
-
-    /// Return `true` if this node is already initialized and cannot be initialized again with
-    /// [`Raft::initialize`]
-    #[since(version = "0.10.0")]
-    pub async fn is_initialized(&self) -> Result<bool, Fatal<C>> {
-        let initialized = self.with_raft_state(|st| st.is_initialized()).await?;
-
-        Ok(initialized)
-    }
-
-    /// Initialize a pristine Raft node with the given config.
-    ///
-    /// This command should be called on pristine nodes — where the log index is 0 and the node is
-    /// in Learner state — as if either of those constraints are false, it indicates that the
-    /// cluster is already formed and in motion. If `InitializeError::NotAllowed` is returned
-    /// from this function, it is safe to ignore, as it simply indicates that the cluster is
-    /// already up and running, which is ultimately the goal of this function. You can check
-    /// if the cluster is initialized with [`Raft::is_initialized()`] and then avoid re-initialize
-    /// it in case you want to get rid of this error.
-    ///
-    /// ## Recommended Usage
-    ///
-    /// The simplest and most appropriate way to initialize a cluster is to call `initialize()`
-    /// on **exactly one node**. The other nodes should remain empty and wait for the initialized
-    /// node to replicate logs to them.
-    ///
-    /// Calling `initialize()` on multiple nodes with **identical configuration** is also
-    /// acceptable and will not cause any consistency issues — the Raft voting protocol ensures
-    /// that only one leader will be elected.
-    ///
-    /// However, calling `initialize()` with **different configurations** on different nodes
-    /// may lead to a split-brain condition and must be avoided.
-    ///
-    /// ## Behavior
-    ///
-    /// Once a node is successfully initialized, it will commit a new membership config
-    /// log entry to store, then enter Candidate state and attempt to elect itself as the
-    /// leader.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// use std::collections::BTreeMap;
-    /// use openraft::BasicNode;
-    ///
-    /// // Initialize a single-node cluster
-    /// let mut nodes = BTreeMap::new();
-    /// nodes.insert(1, BasicNode { addr: "127.0.0.1:8080".to_string() });
-    /// raft.initialize(nodes).await?;
-    ///
-    /// // Initialize a three-node cluster
-    /// let mut nodes = BTreeMap::new();
-    /// nodes.insert(1, BasicNode { addr: "127.0.0.1:8080".to_string() });
-    /// nodes.insert(2, BasicNode { addr: "127.0.0.1:8081".to_string() });
-    /// nodes.insert(3, BasicNode { addr: "127.0.0.1:8082".to_string() });
-    /// raft.initialize(nodes).await?;
-    /// ```
-    #[tracing::instrument(level = "debug", skip(self))]
-    pub async fn initialize<T>(&self, members: T) -> Result<(), RaftError<C, InitializeError<C>>>
-    where T: IntoNodes<C::NodeId, C::Node> + Debug {
-        self.management_api().initialize(members).await.into_raft_result()
-    }
-
-    /// Provides read-only access to [`RaftState`] through a user-provided function.
-    ///
-    /// The function `func` is applied to the current [`RaftState`]. The result of this function,
-    /// of type `V`, is returned wrapped in `Result<V, Fatal<C>>`. `Fatal` error will be
-    /// returned if failed to receive a reply from `RaftCore`.
-    ///
-    /// A `Fatal` error is returned if:
-    /// - Raft core task is stopped normally.
-    /// - Raft core task is panicked due to programming error.
-    /// - Raft core task is encountered a storage error.
-    ///
-    /// Example for getting the current committed log id:
-    /// ```ignore
-    /// let committed = my_raft.with_raft_state(|st| st.committed).await?;
-    /// ```
-    pub async fn with_raft_state<F, V>(&self, func: F) -> Result<V, Fatal<C>>
-    where
-        F: FnOnce(&RaftState<C>) -> V + OptionalSend + 'static,
-        V: OptionalSend + 'static,
-    {
-        let (tx, rx) = C::oneshot();
-
-        self.external_request(|st| {
-            let result = func(st);
-            if let Err(_err) = tx.send(result) {
-                tracing::error!("{}: to-Raft tx send error", func_name!());
-            }
-        })
-        .await?;
-
-        match rx.await {
-            Ok(res) => Ok(res),
-            Err(err) => {
-                tracing::error!("{}: rx recv error: {}", func_name!(), err);
-                let fatal = self.inner.get_core_stop_error().await;
-                Err(fatal)
-            }
-        }
-    }
-
-    /// Send a request to the Raft core loop in a fire-and-forget manner.
-    ///
-    /// This method returns immediately after sending the message to the Raft core loop,
-    /// without waiting for the request to be executed. The returned `Result` indicates
-    /// whether the message was successfully sent, not whether the request was executed.
-    ///
-    /// The request functor will be called with an immutable reference to the [`RaftState`]
-    /// and serialized with other Raft core loop processing (e.g., client requests
-    /// or general state changes).
-    ///
-    /// If a response is required, then the caller can store the sender of a one-shot channel
-    /// in the closure of the request functor, which can then be used to send the response
-    /// asynchronously.
-    ///
-    /// Returns a `Fatal` error if:
-    /// - Raft core task is stopped normally.
-    /// - Raft core task is panicked due to programming error.
-    /// - Raft core task is encountered a storage error.
-    pub async fn external_request<F>(&self, req: F) -> Result<(), Fatal<C>>
-    where F: FnOnce(&RaftState<C>) + OptionalSend + 'static {
-        let req: BoxOnce<'static, RaftState<C>> = Box::new(req);
-        self.inner.send_msg(RaftMsg::WithRaftState { req }).await
-    }
-
-    /// Get a handle to the metrics channel.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// // Get current metrics
-    /// let metrics = raft.metrics().borrow_watched().clone();
-    /// println!("Current leader: {:?}", metrics.current_leader);
-    /// println!("Current term: {}", metrics.current_term);
-    /// ```
-    pub fn metrics(&self) -> WatchReceiverOf<C, RaftMetrics<C>> {
-        self.inner.rx_metrics.clone()
-    }
-
-    /// Get a handle to the data metrics channel.
-    pub fn data_metrics(&self) -> WatchReceiverOf<C, RaftDataMetrics<C>> {
-        self.inner.rx_data_metrics.clone()
-    }
-
-    /// Get a handle to the server metrics channel.
-    pub fn server_metrics(&self) -> WatchReceiverOf<C, RaftServerMetrics<C>> {
-        self.inner.rx_server_metrics.clone()
-    }
-
-    /// Get a handle to watch log I/O flush progress.
-    ///
-    /// Tracks when log entries and votes are durably written to storage.
-    /// Updated on every I/O completion (vote saves and log appends).
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let mut log_progress = raft.watch_log_progress();
-    ///
-    /// // Wait for a specific log entry to be flushed
-    /// let target = Some(FlushPoint::new(
-    ///     Vote::new_committed(2, node_id),
-    ///     Some(LogId::new(LeaderId::new(2, node_id), 100))
-    /// ));
-    /// log_progress.wait_until_ge(&target).await?;
-    /// ```
-    #[since(version = "0.10.0")]
-    #[must_use = "progress handle should be stored to track I/O progress"]
-    pub fn watch_log_progress(&self) -> LogProgress<C> {
-        self.inner.progress_watcher.log_progress()
-    }
-
-    /// Get a handle to watch vote I/O flush progress.
-    ///
-    /// Tracks when votes (leadership changes) are durably written to storage.
-    /// Updated only when the vote changes (new term or leader), not on every log append.
-    ///
-    /// Use this when you only care about leadership changes, not specific log entries.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let mut vote_progress = raft.watch_vote_progress();
-    ///
-    /// // Wait for term 2 to be persisted
-    /// let target = Some(Vote::new_committed(2, 0));
-    /// vote_progress.wait_until_ge(&target).await?;
-    /// ```
-    #[since(version = "0.10.0")]
-    #[must_use = "progress handle should be stored to track vote progress"]
-    pub fn watch_vote_progress(&self) -> VoteProgress<C> {
-        self.inner.progress_watcher.vote_progress()
-    }
-
-    /// Get a handle to watch commit log progress.
-    ///
-    /// Tracks when committed logs advance(persisted on a quorum and the last-log is proposed by the
-    /// leader). Updated whenever the committed cursor moves forward.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let mut commit_progress = raft.watch_commit_progress();
-    ///
-    /// // Wait until log index 42 is committed
-    /// let target = Some(LogId::new(LeaderId::new(2, node_id), 42));
-    /// commit_progress.wait_until_ge(&target).await?;
-    /// ```
-    #[since(version = "0.10.0")]
-    #[must_use = "progress handle should be stored to track commit progress"]
-    pub fn watch_commit_progress(&self) -> CommitProgress<C> {
-        self.inner.progress_watcher.commit_progress()
-    }
-
-    /// Get a handle to watch snapshot persistence progress.
-    ///
-    /// Tracks when snapshots are persisted to storage.
-    /// Updated whenever a snapshot is built or installed and persisted.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let mut snapshot_progress = raft.watch_snapshot_progress();
-    ///
-    /// // Wait until snapshot covering log index 100 is persisted
-    /// let target = Some(LogId::new(LeaderId::new(2, node_id), 100));
-    /// snapshot_progress.wait_until_ge(&target).await?;
-    /// ```
-    #[since(version = "0.10.0")]
-    #[must_use = "progress handle should be stored to track snapshot progress"]
-    pub fn watch_snapshot_progress(&self) -> SnapshotProgress<C> {
-        self.inner.progress_watcher.snapshot_progress()
-    }
-
-    /// Get a handle to watch applied log progress.
-    ///
-    /// Tracks when logs are applied to the state machine.
-    /// Updated whenever the last applied log id advances.
-    ///
-    /// # Note
-    ///
-    /// If the state machine does not persist the applied state immediately, the watcher
-    /// may observe duplicate events when the server restarts and re-applies log entries.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let mut apply_progress = raft.watch_apply_progress();
-    ///
-    /// // Wait until log index 42 is applied
-    /// let target = Some(LogId::new(LeaderId::new(2, node_id), 42));
-    /// apply_progress.wait_until_ge(&target).await?;
-    /// ```
-    #[since(version = "0.10.0")]
-    #[must_use = "progress handle should be stored to track applied progress"]
-    pub fn watch_apply_progress(&self) -> AppliedProgress<C> {
-        self.inner.progress_watcher.apply_progress()
-    }
-
-    /// Watch for any leader changes in the cluster and invoke callback on each change.
-    ///
-    /// Returns a [`WatchChangeHandle`] that must be held to keep watching.
-    /// If the handle is dropped or [`WatchChangeHandle::close()`] is called,
-    /// the background task will be terminated and the callback will no longer be invoked.
-    ///
-    /// The callback receives:
-    /// - `old`: The previous leader state `(leader_id, committed)`, or `None` on the first callback
-    /// - `new`: The current leader state `(leader_id, committed)`
-    ///
-    /// This fires on ANY leader change in the cluster, not just when this node's leadership
-    /// changes. For a simpler API that only fires when THIS node becomes or loses leadership,
-    /// see [`on_leader_change()`].
-    ///
-    /// # Note on Start/Stop Service Pattern
-    ///
-    /// If you use this API to start/stop services based on leadership, be aware that
-    /// consecutive callbacks may show the same node as leader with different Terms
-    /// (e.g., Term 1 → Term 2). The simple `if is_leader { start } else { stop }` pattern
-    /// could call `start` twice without an intervening `stop`.
-    ///
-    /// For the start/stop service pattern, prefer [`on_leader_change()`] which guarantees
-    /// alternating `start`/`stop` callbacks.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let my_node_id = 1;
-    ///
-    /// let mut handle = raft.on_cluster_leader_change(move |_old, (leader_id, committed)| {
-    ///     let is_leader = leader_id.node_id == my_node_id && committed;
-    ///
-    ///     async move {
-    ///         if is_leader {
-    ///             // This node just became the committed leader
-    ///             // Start leader-only services (e.g., cron jobs, cache warming)
-    ///             start_leader_services().await;
-    ///         } else {
-    ///             // This node is no longer the leader
-    ///             // Stop leader-only services to avoid duplicate work
-    ///             stop_leader_services().await;
-    ///         }
-    ///     }
-    /// });
-    ///
-    /// // Later, stop watching
-    /// handle.close().await;
-    /// ```
-    ///
-    /// [`on_leader_change()`]: Self::on_leader_change
-    #[since(version = "0.10.0")]
-    #[must_use = "handle must be held to keep the watch task running"]
-    pub fn on_cluster_leader_change<F, Fut>(&self, mut callback: F) -> WatchChangeHandle<C>
-    where
-        F: FnMut(Option<(C::LeaderId, bool)>, (C::LeaderId, bool)) -> Fut + OptionalSend + 'static,
-        Fut: Future<Output = ()> + OptionalSend + 'static,
-    {
-        let mut prev_vote: Option<Vote<C::LeaderId>> = None;
-
-        self.watch_vote_change(move |new_vote, _my_node_id| {
-            let old_leader = prev_vote.as_ref().map(|v| v.leader_id().clone());
-            let new_leader = new_vote.leader_id().clone();
-
-            // Only call callback if leader_id actually changed
-            let fut = if old_leader.as_ref() != Some(&new_leader) {
-                let old_state = prev_vote.as_ref().map(|v| (v.leader_id().clone(), v.is_committed()));
-                let new_state = (new_vote.leader_id().clone(), new_vote.is_committed());
-                Some(callback(old_state, new_state))
-            } else {
-                None
-            };
-            prev_vote = Some(new_vote);
-
-            async move {
-                if let Some(f) = fut {
-                    f.await;
-                }
-            }
-        })
-    }
-
-    /// Register callbacks for when this node becomes or stops being the committed leader.
-    ///
-    /// Returns a [`WatchChangeHandle`] that must be held to keep watching.
-    /// If the handle is dropped or [`WatchChangeHandle::close()`] is called,
-    /// the background task will be terminated and the callbacks will no longer be invoked.
-    ///
-    /// Unlike [`on_cluster_leader_change()`] which fires on any leader change in the cluster,
-    /// this method only fires when THIS node becomes or stops being the leader.
-    ///
-    /// - `start`: Called when this node becomes the leader (committed, quorum-acknowledged)
-    /// - `stop`: Called when this node is no longer the leader (another node becomes leader)
-    ///
-    /// # Callback Guarantees
-    ///
-    /// The `start` and `stop` callbacks are guaranteed to be called in alternating order:
-    /// `start` → `stop` → `start` → `stop` → ...
-    ///
-    /// Even if a node transitions directly from leader in Term 1 to leader in Term 2,
-    /// `stop` will be called with the old `leader_id` before `start` is called with the
-    /// new `leader_id`. This ensures proper resource cleanup between leadership terms.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// let handle = raft.on_leader_change(
-    ///     |leader_id| async move {
-    ///         println!("Became leader: {:?}", leader_id);
-    ///         start_leader_services().await;
-    ///     },
-    ///     |old_leader| async move {
-    ///         println!("Stopped leading: {:?}", old_leader);
-    ///         stop_leader_services().await;
-    ///     },
-    /// );
-    ///
-    /// // Later, stop watching
-    /// handle.close().await;
-    /// ```
-    ///
-    /// [`on_cluster_leader_change()`]: Self::on_cluster_leader_change
-    #[since(version = "0.10.0")]
-    #[must_use = "handle must be held to keep the watch task running"]
-    pub fn on_leader_change<F1, F2, Fut1, Fut2>(&self, start: F1, stop: F2) -> WatchChangeHandle<C>
-    where
-        F1: Fn(C::LeaderId) -> Fut1 + OptionalSend + 'static,
-        F2: Fn(C::LeaderId) -> Fut2 + OptionalSend + 'static,
-        Fut1: Future<Output = ()> + OptionalSend + 'static,
-        Fut2: Future<Output = ()> + OptionalSend + 'static,
-    {
-        let mut prev_leader_id = None;
-
-        self.watch_vote_change(move |vote, my_node_id| {
-            let leader_id = vote.leader_id().clone();
-
-            // Fire `start` when THIS node becomes committed leader
-            // and it's a new leadership (different from current)
-            #[allow(clippy::collapsible_else_if)]
-            let (stop_fut, start_fut) = if leader_id.node_id() == my_node_id {
-                if vote.is_committed() && prev_leader_id.as_ref() != Some(&leader_id) {
-                    // Call stop first if transitioning from one leadership to another
-                    // (e.g., Term 1 leader -> Term 2 leader)
-                    // This guarantees alternating start/stop calls.
-                    let stop_fut = prev_leader_id.take().map(&stop);
-                    let start_fut = Some(start(leader_id.clone()));
-                    prev_leader_id = Some(leader_id);
-                    (stop_fut, start_fut)
-                } else {
-                    (None, None)
-                }
-            } else {
-                let stop_fut = prev_leader_id.take().map(&stop);
-                (stop_fut, None)
-            };
-
-            async move {
-                if let Some(f) = stop_fut {
-                    f.await;
-                }
-                if let Some(f) = start_fut {
-                    f.await;
-                }
-            }
-        })
-    }
-
-    /// Spawn a task that watches vote changes and invokes async callback on each change.
-    ///
-    /// This is an internal helper used by [`Self::on_leader_change()`] and
-    /// [`Self::on_cluster_leader_change()`].
-    ///
-    /// The callback returns a future that will be awaited before processing
-    /// the next vote change.
-    fn watch_vote_change<F, Fut>(&self, mut callback: F) -> WatchChangeHandle<C>
-    where
-        F: FnMut(Vote<C::LeaderId>, &NodeIdOf<C>) -> Fut + OptionalSend + 'static,
-        Fut: Future<Output = ()> + OptionalSend + 'static,
-    {
-        use futures_util::FutureExt;
-
-        let my_node_id = self.inner.id().clone();
-        let mut vote_progress = self.watch_vote_progress();
-        let (cancel_tx, cancel_rx) = C::oneshot::<()>();
-
-        let handle = C::spawn(async move {
-            let mut cancel_rx = cancel_rx.fuse();
-
-            loop {
-                futures_util::select! {
-                    _ = cancel_rx => break,
-                    res = vote_progress.changed().fuse() => {
-                        if res.is_err() {
-                            break;
-                        }
-                        let Some(vote) = vote_progress.get() else {
-                            continue;
-                        };
-
-                        callback(vote, &my_node_id).await;
-                    }
-                }
-            }
-        });
-
-        WatchChangeHandle {
-            cancel_tx: Some(cancel_tx),
-            join_handle: Some(handle),
-        }
-    }
-
-    /// Get a handle to wait for the metrics to satisfy some condition.
-    ///
-    ///
-    /// If `timeout` is `None`, then it will wait forever(10 years).
-    /// If `timeout` is `Some`, then it will wait for the specified duration.
-    ///
-    /// ```ignore
-    /// # use std::time::Duration;
-    /// # use openraft::{State, Raft};
-    ///
-    /// let timeout = Duration::from_millis(200);
-    ///
-    /// // wait for raft log-3 to be received and applied:
-    /// r.wait(Some(timeout)).log(Some(3), "log").await?;
-    ///
-    /// // wait for ever for raft node's current leader to become 3:
-    /// r.wait(None).current_leader(2, "wait for leader").await?;
-    ///
-    /// // wait for raft state to become a follower
-    /// r.wait(None).state(State::Follower, "state").await?;
-    /// ```
-    pub fn wait(&self, timeout: Option<Duration>) -> Wait<C> {
-        self.inner.wait(timeout)
-    }
-
-    /// Wait for this node's state machine to recover to at least the state it had before restart.
-    ///
-    /// Call this **once**, right after the node is created, before it starts serving requests. It
-    /// is a one-time step for recovering across a restart, **not** a per-read primitive: it does
-    /// not make later reads linearizable — use [`ensure_linearizable`](Self::ensure_linearizable)
-    /// for each linearizable read instead. (Once recovered, the node stays recovered, so calling it
-    /// again is a no-op that returns as soon as the current cluster commit has been applied.)
-    ///
-    /// The method waits in two phases, together bounded by `timeout`:
-    /// 1. until `cluster_committed` is perceived as covering this node's current durable log tail
-    ///    (the cluster commit is re-established for everything this node may have applied),
-    /// 2. until the state machine has applied up to that cluster commit.
-    ///
-    /// The target is pinned to the first `cluster_committed` that covers this node's current
-    /// durable log tail. If the cluster advanced while this node was down, that target may exceed
-    /// the node's own pre-restart applied index — which still satisfies "at least the same state".
-    ///
-    /// It requires a reachable, functioning cluster: if no cluster commit is re-established (no
-    /// leader, lost quorum, or this node is isolated), the method returns [`WaitError::Timeout`].
-    /// If `timeout` is `None` it waits forever (see [`wait`](Self::wait)).
-    ///
-    /// # Why this works
-    ///
-    /// Perceiving a `cluster_committed` that covers this node's durable log tail means the current
-    /// leader has either committed the local tail this node may have applied before restart, or has
-    /// first replaced any uncommitted tail with the leader's log. Applying up to that commit
-    /// restores every durable log entry that could have contributed to the pre-restart state. This
-    /// is the same `read_log_id` argument that makes linearizable reads safe: see [the read
-    /// protocol](crate::docs::protocol::read) for the read-safety argument and [the commit
-    /// protocol](crate::docs::protocol::commit) for why `cluster_committed` is never a value
-    /// restored from storage. A node that did not persist `committed` can use this in place of, or
-    /// together with, saving it (see [log pointers](crate::docs::data::log_pointers)).
-    ///
-    /// # Limitation: a follower may still recover a stale state
-    ///
-    /// The soundness above is bounded by what this node made **durable**. A follower may apply an
-    /// entry the cluster has committed before it flushes that entry locally — apply is gated on the
-    /// cluster commit, not on the local flush (the invariant is `applied ≤ committed ≤ submitted`,
-    /// not `applied ≤ flushed`; see [log pointers](crate::docs::data::log_pointers)). If the node
-    /// restarts while such an entry is applied but still unflushed, and its applied state was not
-    /// otherwise persisted, the restart reverts three tracking points at once: the durable log tail
-    /// falls back to `flushed`, the state machine falls back to an earlier applied state, and
-    /// `last_log_index` now describes the shorter log.
-    ///
-    /// Because the first-phase wait is `cluster_committed >= last_log_index` and `last_log_index`
-    /// has itself reverted, a stale `cluster_committed` covering only the shrunken log satisfies
-    /// it. `wait_for_recovery` can then return having recovered a state older than the one this
-    /// node served immediately before the restart. The node is not stuck — the leader
-    /// re-replicates the lost tail and it catches up — but across that window the method no
-    /// longer pins recovery to the pre-restart state.
-    ///
-    /// This can only happen while serving reads as a follower. A leader holds every committed entry
-    /// and never applies below its commit, so a read taken on the leader through
-    /// [`ensure_linearizable`](Self::ensure_linearizable) cannot revert to an earlier state. When a
-    /// read must be guaranteed fresh across a restart, serve it on the leader.
-    ///
-    /// # Example
-    ///
-    /// ```ignore
-    /// // On startup, block until the state machine has recovered the committed state before
-    /// // serving reads:
-    /// raft.wait_for_recovery(Some(Duration::from_secs(5))).await?;
-    /// ```
-    #[since(version = "0.10.0")]
-    pub async fn wait_for_recovery(&self, timeout: Option<Duration>) -> Result<RaftMetrics<C>, WaitError> {
-        let start = C::now();
-
-        // Phase 1: wait until the cluster commit is re-established for this node's durable log
-        // tail. A follower may first hear an older leader_commit, which is not enough to recover
-        // everything it had applied before restart.
-        let metrics = self
-            .wait(timeout)
-            .metrics(
-                |m| {
-                    let cluster_committed = m.cluster_committed.as_ref().map(|x| x.index());
-                    cluster_committed.is_some() && cluster_committed >= m.last_log_index
-                },
-                "wait_for_recovery: cluster commit covers local log",
-            )
-            .await?;
-
-        let target = metrics.cluster_committed.as_ref().map(|x| x.index());
-
-        // Phase 2: wait until the state machine has applied up to that cluster commit, within the
-        // time remaining from the original budget.
-        let remaining = timeout.map(|t| t.saturating_sub(C::now() - start));
-        self.wait(remaining)
-            .applied_index_at_least(target, "wait_for_recovery: state machine applied cluster commit")
-            .await
-    }
-
-    /// Shutdown this Raft node.
-    ///
-    /// It sends a shutdown signal and waits until `RaftCore` returns.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// // Gracefully shutdown the Raft node
-    /// raft.shutdown().await?;
-    /// ```
-    pub async fn shutdown(&self) -> Result<(), JoinErrorOf<C>> {
-        if let Some(tx) = self.inner.tx_shutdown.lock().unwrap().take() {
-            // A failure to send means the RaftCore is already shutdown. Continue to check the task
-            // return value.
-            let send_res = tx.send(());
-            tracing::info!("sending shutdown signal to RaftCore, sending res: {:?}", send_res);
-        }
-        self.inner.join_core_task().await;
-        if let Some(join_handle) = self.inner.tick_handle.shutdown() {
-            join_handle.await.ok();
-        }
-
-        // TODO(xp): API change: replace `JoinError` with `Fatal`,
-        //           to let the caller know the return value of RaftCore task.
-        Ok(())
-    }
-
-    /// Provides mutable access to [`RaftStateMachine`] through a user-provided function.
-    ///
-    /// The function `func` is applied to the current [`RaftStateMachine`]. The result of this
-    /// function, of type `V`, is returned wrapped in `Result<V, Fatal<C>>`.
-    ///
-    /// A `Fatal` error is returned if:
-    /// - Raft core task is stopped normally.
-    /// - Raft core task is panicked due to programming error.
-    /// - Raft core task is encountered a storage error.
-    ///
-    /// Example for getting the last applied log id from SM(assume there is `last_applied()` method
-    /// provided):
-    ///
-    /// ```rust,ignore
-    /// let last_applied_log_id = my_raft.with_state_machine(|sm| {
-    ///     async move { sm.last_applied().await }
-    /// }).await?;
-    /// ```
-    #[since(version = "0.10.0")]
-    pub async fn with_state_machine<F, V>(&self, func: F) -> Result<V, Fatal<C>>
-    where
-        SM: OptionalSend + 'static,
-        F: FnOnce(&mut SM) -> BoxFuture<V> + OptionalSend + 'static,
-        V: OptionalSend + 'static,
-    {
-        let (tx, rx) = C::oneshot();
-
-        self.external_state_machine_request(|sm| {
-            Box::pin(async move {
-                let resp = func(sm).await;
-                if let Err(_err) = tx.send(resp) {
-                    tracing::error!("{}: failed to send response to user tx", func_name!());
-                }
-            })
-        })
-        .await?;
-
-        // Use the bounded receive: the state-machine worker owns the responder and can die on its
-        // own (e.g. a panic in `func`) while RaftCore keeps running. An unbounded join on the core
-        // would then hang forever. `recv_msg` waits only up to `RECV_CORE_STOP_TIMEOUT` before
-        // resolving to `Fatal::Stopped`.
-        self.inner.recv_msg(rx).await
-    }
-
-    /// Send a request to the [`RaftStateMachine`] worker in a fire-and-forget manner.
-    ///
-    /// This method returns immediately after sending the message to the state machine worker,
-    /// without waiting for the request to be executed. The returned `Result` indicates
-    /// whether the message was successfully sent, not whether the request was executed.
-    ///
-    /// The request functor will be called with a mutable reference to the state machine.
-    /// The functor returns a [`Future`] because state machine methods are `async`.
-    ///
-    /// Returns a `Fatal` error if:
-    /// - Raft core task is stopped normally.
-    /// - Raft core task is panicked due to programming error.
-    /// - Raft core task is encountered a storage error.
-    #[since(version = "0.10.0")]
-    pub async fn external_state_machine_request<F>(&self, req: F) -> Result<(), Fatal<C>>
-    where
-        SM: OptionalSend + 'static,
-        F: FnOnce(&mut SM) -> BoxFuture<()> + OptionalSend + 'static,
-    {
-        let Some(tx) = self.sm_cmd_tx.upgrade() else {
-            return Err(self.inner.get_core_stop_error_bounded().await);
-        };
-
-        let sm_cmd = sm::Command::ExternalFunc {
-            func: Box::new(move |sm| req(sm)),
-        };
-        if tx.send(sm_cmd).await.is_err() {
-            return Err(self.inner.get_core_stop_error_bounded().await);
-        }
-        Ok(())
     }
 }
